Rename trie to utf8Trie
